IDP-1838 Pooled methods to avoid duplicate code
